:shirt: Move converters to Format namespace
