+++ src/Texim/Formats/IndexedImageBitmapParameters.cs
@@ -17,42 +17,17 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-namespace Texim.Images
+namespace Texim.Formats
 {
-    using System;
-    using System.Drawing;
     using System.Drawing.Imaging;
-    using Yarhl.FileFormat;
-    using Yarhl.IO;
+    using Texim.Palettes;
 
-    public class FullImage2BinaryBitmap :
-        IInitializer<ImageFormat>, IConverter<IFullImage, BinaryFormat>
+    public class IndexedImageBitmapParameters
     {
-        private ImageFormat format = ImageFormat.Png;
+        public ImageFormat Format { get; set; }
 
-        public void Initialize(ImageFormat parameters)
-        {
-            if (parameters == null)
-                throw new ArgumentNullException(nameof(parameters));
+        public IPalette Palette { get; set; }
 
-            format = parameters;
-        }
-
-        public BinaryFormat Convert(IFullImage source)
-        {
-            if (source == null)
-                throw new ArgumentNullException(nameof(source));
-
-            var image = new Bitmap(source.Width, source.Height);
-            for (int x = 0; x < source.Width; x++) {
-                for (int y = 0; y < source.Height; y++) {
-                    image.SetPixel(x, y, source.Pixels[y * source.Width + x].ToColor());
-                }
-            }
-
-            var binary = new BinaryFormat();
-            image.Save(binary.Stream, format);
-            return binary;
-        }
+        public IPaletteCollection Palettes { get; set; }
     }
 }
